+++ tee-worker/litentry/core/stf-task/sender/src/error.rs
@@ -16,6 +16,7 @@
 
 #[cfg(all(not(feature = "std"), feature = "sgx"))]
 use crate::sgx_reexport_prelude::*;
+
 use std::boxed::Box;
 
 pub type Result<T> = core::result::Result<T, Error>;
